Adding confs to setup VIrtaul Lab

+++ hashicorp.sh
@@ -1,5 +1,15 @@
-# TERRAFORM
+# HASHICORP REPO
+if [ ! -f /etc/apt/sources.list.d/hashicorp.list ]; then
+    wget -O- https://apt.releases.hashicorp.com/gpg | sudo gpg --dearmor -o /usr/share/keyrings/hashicorp-archive-keyring.gpg
+    echo "deb [signed-by=/usr/share/keyrings/hashicorp-archive-keyring.gpg] https://apt.releases.hashicorp.com $(lsb_release -cs) main" | sudo tee /etc/apt/sources.list.d/hashicorp.list
+fi
 
-wget -O- https://apt.releases.hashicorp.com/gpg | sudo gpg --dearmor -o /usr/share/keyrings/hashicorp-archive-keyring.gpg
-echo "deb [signed-by=/usr/share/keyrings/hashicorp-archive-keyring.gpg] https://apt.releases.hashicorp.com $(lsb_release -cs) main" | sudo tee /etc/apt/sources.list.d/hashicorp.list
-sudo apt update && sudo apt install terraform
+# TERRAFORM
+if ! terraform -v; then
+    sudo apt update && sudo apt install terraform
+fi
+
+# VAGRANT
+if ! vagrant -v; then
+    sudo apt update && sudo apt install vagrant
+fi
